Remove buildSrc
