--- buildSrc/src/main/kotlin/BuildPlugins.kt
@@ -1,4 +1,0 @@
-object BuildPlugins {
-    const val KTLINT = "org.jlleitschuh.gradle.ktlint"
-    const val PAPARAZZI = "app.cash.paparazzi"
-}
