--- buildSrc/src/main/kotlin/extensions/RepositoryHandlerExtensions.kt
@@ -1,11 +1,0 @@
-package extensions
-
-import org.gradle.api.artifacts.dsl.RepositoryHandler
-import org.gradle.kotlin.dsl.maven
-
-fun RepositoryHandler.applyDefault() {
-    google()
-    mavenCentral()
-    maven("https://plugins.gradle.org/m2/")
-    maven("https://customers.pspdfkit.com/maven")
-}
