+++ common/design-compose/build.gradle.kts
@@ -27,10 +27,10 @@ plugins {
 }
 
 android {
-    compileSdk = BuildAndroidConfig.COMPILE_SDK_VERSION
+    compileSdk = 33
 
     defaultConfig {
-        minSdk = BuildAndroidConfig.MIN_SDK_VERSION
+        minSdk = 21
     }
 
     namespace = "app.ss.design.compose"
