+++ common/storage/build.gradle.kts
@@ -28,10 +28,10 @@ plugins {
 }
 
 android {
-    compileSdk = BuildAndroidConfig.COMPILE_SDK_VERSION
+    compileSdk = 33
 
     defaultConfig {
-        minSdk = BuildAndroidConfig.MIN_SDK_VERSION
+        minSdk = 21
 
         javaCompileOptions {
             annotationProcessorOptions {
