+++ features/lessons/build.gradle.kts
@@ -30,10 +30,10 @@ plugins {
 }
 
 android {
-    compileSdk = BuildAndroidConfig.COMPILE_SDK_VERSION
+    compileSdk = 33
 
     defaultConfig {
-        minSdk = BuildAndroidConfig.MIN_SDK_VERSION
+        minSdk = 21
     }
 
     namespace = "com.cryart.sabbathschool.lessons"
